requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
apscheduler==3.10.4
beautifulsoup4==4.12.3
tenacity==9.0.0
aiohttp==3.10.5
python-dotenv==1.0.1
aiosqlite==0.20.0
feedparser>=6.0
trafilatura>=1.6
readability-lxml>=0.4
jieba>=0.42
filelock>=3.12
